kuika-flask/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
pycryptodome==3.20.0
PyJWT==2.8.0
pymongo==4.8.0
Werkzeug==3.0.3
